feat: move response constructor functions to utils.go
